Move base into v04 (#83)
* Move base into v04

* Add base to docs

* pre-commit fixes

* Fix type_checking block
