+++ src/ome_zarr_models/_utils.py
@@ -1,3 +1,7 @@
+"""
+Private utilities.
+"""
+
 from collections import Counter
 from collections.abc import Hashable, Iterable
 from dataclasses import MISSING, fields, is_dataclass
