+++ src/ome_zarr_models/v04/image_label.py
@@ -12,7 +12,7 @@ from pydantic import AfterValidator, Field, model_validator
 from pydantic_zarr.v2 import ArraySpec, GroupSpec
 
 from ome_zarr_models._utils import duplicates
-from ome_zarr_models.base import Base
+from ome_zarr_models.v04.base import Base
 from ome_zarr_models.v04.image import Image, _check_arrays_compatible
 from ome_zarr_models.v04.multiscales import Multiscale
